# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chowdhary Mart API — hyperlocal e-commerce platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove cart item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"

  # ── Wishlist ──────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List my orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate and apply a coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Coupon valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Wallet ────────────────────────────────────────────────────────────────
...
  /admin/coupons:
    get:
      operationId: listAdminCoupons
      tags: [admin]
      summary: List all coupons (admin)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [admin]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
...
components:
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Store:
      type: object
      required: [id, name, lat, lng, address]
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        address:
          type: string
        city:
          type: ["string", "null"]
        rating:
          type: ["string", "null"]
        ratingCount:
          type: ["integer", "null"]
        isOpen:
          type: boolean
        estimatedDeliveryMins:
          type: ["integer", "null"]
        deliveryFee:
          type: ["string", "null"]
        freeDeliveryAbove:
          type: ["string", "null"]
        minOrderValue:
          type: ["string", "null"]
        lat:
          type: number
        lng:
          type: number
...
    Product:
      type: object
      required: [id, name, price, mrp]
      properties:
        id:
          type: integer
        storeId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        mrp:
          type: string
        discountPercent:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        weight:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        stock:
          type: integer
        rating:
          type: ["string", "null"]
        reviewCount:
          type: ["integer", "null"]
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          type: string
...
    CartItemDetail:
      type: object
      required: [id, productId, qty, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        qty:
          type: integer
        price:
          type: string
        product:
          $ref: "#/components/schemas/Product"
    CartResponse:
      type: object
      required: [items, subtotal, deliveryFee, total]
      properties:
        storeId:
          type: ["integer", "null"]
        store:
          $ref: "#/components/schemas/Store"
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemDetail"
        subtotal:
          type: string
        deliveryFee:
          type: string
        discount:
          type: string
        total:
          type: string
        itemCount:
          type: integer
    CartItemInput:
      type: object
      required: [productId, qty]
      properties:
        productId:
          type: integer
        qty:
          type: integer
    CartItemUpdate:
      type: object
      required: [qty]
      properties:
        qty:
          type: integer

    # ── Wishlist ───────────────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, orderNumber, status, total, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: integer
        storeId:
          type: integer
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        subtotal:
          type: string
        deliveryFee:
          type: string
        discount:
          type: string
        couponCode:
          type: ["string", "null"]
        couponDiscount:
          type: ["string", "null"]
        walletUsed:
          type: ["string", "null"]
        total:
          type: string
        loyaltyPointsEarned:
          type: ["integer", "null"]
        estimatedDeliveryMins:
          type: ["integer", "null"]
        createdAt:
          type: string
        store:
          $ref: "#/components/schemas/Store"
...
    OrderInput:
      type: object
      required: [addressId, paymentMethod]
      properties:
        addressId:
          type: integer
        paymentMethod:
          type: string
          enum: [cod, online, wallet, upi]
        couponCode:
          type: string
        useWallet:
          type: boolean
        notes:
          type: string
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue]
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
        discountValue:
          type: string
        minOrderValue:
          type: ["string", "null"]
        maxDiscount:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [percent, flat]
        discountValue:
          type: string
        minOrderValue:
          type: string
        maxDiscount:
          type: string
        usageLimit:
          type: integer
        perUserLimit:
          type: integer
        expiresAt:
          type: string
    CouponValidateInput:
      type: object
      required: [code, orderValue]
      properties:
        code:
          type: string
        orderValue:
          type: string
    CouponValidateResponse:
      type: object
      required: [valid, discount]
      properties:
        valid:
          type: boolean
        discount:
          type: string
        message:
          type: ["string", "null"]
        coupon:
          $ref: "#/components/schemas/Coupon"

    # ── Wallet ─────────────────────────────────────────────────────────────
...